Cleanup Part 2: Electric Boogaloo (Reloaded)

+++ src/client/rest/RESTMethods.js
@@ -1,11 +1,11 @@
 const Constants = require('../../util/Constants');
 const Collection = require('../../util/Collection');
 
-const getStructure = name => require(`../../structures/${name}`);
-const User = getStructure('User');
-const GuildMember = getStructure('GuildMember');
-const Role = getStructure('Role');
-const Invite = getStructure('Invite');
+const requireStructure = name => require(`../../structures/${name}`);
+const User = requireStructure('User');
+const GuildMember = requireStructure('GuildMember');
+const Role = requireStructure('Role');
+const Invite = requireStructure('Invite');
 
 class RESTMethods {
   constructor(restManager) {
@@ -45,10 +45,9 @@ class RESTMethods {
     });
   }
 
-  sendMessage($channel, content, tts, nonce, file) {
+  sendMessage(channel, content, tts, nonce, file) {
     return new Promise((resolve, reject) => {
       const $this = this;
-      let channel = $channel;
 
       function req() {
         $this.rest.makeRequest('post', Constants.Endpoints.channelMessages(channel.id), true, {
@@ -101,11 +100,9 @@ class RESTMethods {
     return new Promise((resolve, reject) => {
       this.rest.makeRequest('patch', Constants.Endpoints.channelMessage(message.channel.id, message.id), true, {
         content,
-      })
-      .then(data => {
+      }).then(data => {
         resolve(this.rest.client.actions.MessageUpdate.handle(data).updated);
-      })
-      .catch(reject);
+      }).catch(reject);
     });
   }
 
@@ -114,11 +111,9 @@ class RESTMethods {
       this.rest.makeRequest('post', Constants.Endpoints.guildChannels(guild.id), true, {
         name: channelName,
         type: channelType,
-      })
-        .then(data => {
-          resolve(this.rest.client.actions.ChannelCreate.handle(data).channel);
-        })
-        .catch(reject);
+      }).then(data => {
+        resolve(this.rest.client.actions.ChannelCreate.handle(data).channel);
+      }).catch(reject);
     });
   }
 
@@ -126,98 +121,73 @@ class RESTMethods {
     const dmChannel = Array.from(this.rest.client.channels.values())
       .filter(channel => channel.recipient)
       .filter(channel => channel.recipient.id === recipient.id);
-
     return dmChannel[0];
   }
 
   createDM(recipient) {
     return new Promise((resolve, reject) => {
       const dmChannel = this.getExistingDM(recipient);
-
-      if (dmChannel) {
-        return resolve(dmChannel);
-      }
-
+      if (dmChannel) return resolve(dmChannel);
       return this.rest.makeRequest('post', Constants.Endpoints.userChannels(this.rest.client.user.id), true, {
         recipient_id: recipient.id,
-      })
-      .then(data => resolve(this.rest.client.actions.ChannelCreate.handle(data).channel))
-      .catch(reject);
+      }).then(data => resolve(this.rest.client.actions.ChannelCreate.handle(data).channel)).catch(reject);
     });
   }
 
-  deleteChannel($channel) {
-    let channel = $channel;
+  deleteChannel(channel) {
     return new Promise((resolve, reject) => {
-      if (channel instanceof User || channel instanceof GuildMember) {
-        channel = this.getExistingDM(channel);
-      }
-
-      this.rest.makeRequest('del', Constants.Endpoints.channel(channel.id), true)
-        .then($data => {
-          const data = $data;
-          data.id = channel.id;
-          resolve(this.rest.client.actions.ChannelDelete.handle(data).channel);
-        })
-        .catch(reject);
+      if (channel instanceof User || channel instanceof GuildMember) channel = this.getExistingDM(channel);
+      this.rest.makeRequest('del', Constants.Endpoints.channel(channel.id), true).then(data => {
+        data.id = channel.id;
+        resolve(this.rest.client.actions.ChannelDelete.handle(data).channel);
+      }).catch(reject);
     });
   }
 
-  updateChannel(channel, $data) {
-    const data = $data;
+  updateChannel(channel, data) {
     return new Promise((resolve, reject) => {
       data.name = (data.name || channel.name).trim();
       data.topic = data.topic || channel.topic;
       data.position = data.position || channel.position;
       data.bitrate = data.bitrate || channel.bitrate;
 
-      this.rest.makeRequest('patch', Constants.Endpoints.channel(channel.id), true, data)
-        .then(newData => {
-          resolve(this.rest.client.actions.ChannelUpdate.handle(newData).updated);
-        })
-        .catch(reject);
+      this.rest.makeRequest('patch', Constants.Endpoints.channel(channel.id), true, data).then(newData => {
+        resolve(this.rest.client.actions.ChannelUpdate.handle(newData).updated);
+      }).catch(reject);
     });
   }
 
   leaveGuild(guild) {
-    if (guild.ownerID === this.rest.client.user.id) {
-      return this.deleteGuild(guild);
-    }
+    if (guild.ownerID === this.rest.client.user.id) return this.deleteGuild(guild);
     return new Promise((resolve, reject) => {
-      this.rest.makeRequest('del', Constants.Endpoints.meGuild(guild.id), true)
-        .then(() => {
-          resolve(this.rest.client.actions.GuildDelete.handle({ id: guild.id }).guild);
-        })
-        .catch(reject);
+      this.rest.makeRequest('del', Constants.Endpoints.meGuild(guild.id), true).then(() => {
+        resolve(this.rest.client.actions.GuildDelete.handle({ id: guild.id }).guild);
+      }).catch(reject);
     });
   }
 
   // untested but probably will work
   deleteGuild(guild) {
     return new Promise((resolve, reject) => {
-      this.rest.makeRequest('del', Constants.Endpoints.guild(guild.id), true)
-        .then(() => {
-          resolve(this.rest.client.actions.GuildDelete.handle({ id: guild.id }).guild);
-        })
-        .catch(reject);
+      this.rest.makeRequest('del', Constants.Endpoints.guild(guild.id), true).then(() => {
+        resolve(this.rest.client.actions.GuildDelete.handle({ id: guild.id }).guild);
+      }).catch(reject);
     });
   }
 
   getUser(userID) {
     return new Promise((resolve, reject) => {
-      this.rest.makeRequest('get', Constants.Endpoints.user(userID), true)
-        .then((data) => {
-          resolve(this.rest.client.actions.UserGet.handle(data).user);
-        })
-        .catch(reject);
+      this.rest.makeRequest('get', Constants.Endpoints.user(userID), true).then((data) => {
+        resolve(this.rest.client.actions.UserGet.handle(data).user);
+      }).catch(reject);
     });
   }
 
   updateCurrentUser(_data) {
     return new Promise((resolve, reject) => {
       const user = this.rest.client.user;
-      const data = {};
 
+      const data = {};
       data.username = _data.username || user.username;
       data.avatar = this.rest.client.resolver.resolveBase64(_data.avatar) || user.avatar;
       if (!user.bot) {
@@ -234,44 +204,15 @@ class RESTMethods {
 
   updateGuild(guild, _data) {
     return new Promise((resolve, reject) => {
-      /*
-      	can contain:
-      	name, region, verificationLevel, afkChannel, afkTimeout, icon, owner, splash
-       */
-
       const data = {};
-
-      if (_data.name) {
-        data.name = _data.name;
-      }
-
-      if (_data.region) {
-        data.region = _data.region;
-      }
-
-      if (_data.verificationLevel) {
-        data.verification_level = Number(_data.verificationLevel);
-      }
-
-      if (_data.afkChannel) {
-        data.afk_channel_id = this.rest.client.resolver.resolveChannel(_data.afkChannel).id;
-      }
-
-      if (_data.afkTimeout) {
-        data.afk_timeout = Number(_data.afkTimeout);
-      }
-
-      if (_data.icon) {
-        data.icon = this.rest.client.resolver.resolveBase64(_data.icon);
-      }
-
-      if (_data.owner) {
-        data.owner_id = this.rest.client.resolver.resolveUser(_data.owner).id;
-      }
-
-      if (_data.splash) {
-        data.splash = this.rest.client.resolver.resolveBase64(_data.splash);
-      }
+      if (_data.name) data.name = _data.name;
+      if (_data.region) data.region = _data.region;
+      if (_data.verificationLevel) data.verification_level = Number(_data.verificationLevel);
+      if (_data.afkChannel) data.afk_channel_id = this.rest.client.resolver.resolveChannel(_data.afkChannel).id;
+      if (_data.afkTimeout) data.afk_timeout = Number(_data.afkTimeout);
+      if (_data.icon) data.icon = this.rest.client.resolver.resolveBase64(_data.icon);
+      if (_data.owner) data.owner_id = this.rest.client.resolver.resolveUser(_data.owner).id;
+      if (_data.splash) data.splash = this.rest.client.resolver.resolveBase64(_data.splash);
 
       this.rest.makeRequest('patch', Constants.Endpoints.guild(guild.id), true, data)
         .then(newData => resolve(this.rest.client.actions.GuildUpdate.handle(newData).updated))
@@ -281,40 +222,34 @@ class RESTMethods {
 
   kickGuildMember(guild, member) {
     return new Promise((resolve, reject) => {
-      this.rest.makeRequest('del', Constants.Endpoints.guildMember(guild.id, member.id), true)
-        .then(() => {
-          resolve(this.rest.client.actions.GuildMemberRemove.handle({
-            guild_id: guild.id,
-            user: member.user,
-          }).member);
-        })
-        .catch(reject);
+      this.rest.makeRequest('del', Constants.Endpoints.guildMember(guild.id, member.id), true).then(() => {
+        resolve(this.rest.client.actions.GuildMemberRemove.handle({
+          guild_id: guild.id,
+          user: member.user,
+        }).member);
+      }).catch(reject);
     });
   }
 
   createGuildRole(guild) {
     return new Promise((resolve, reject) => {
-      this.rest.makeRequest('post', Constants.Endpoints.guildRoles(guild.id), true)
-        .then(role => {
-          resolve(this.rest.client.actions.GuildRoleCreate.handle({
-            guild_id: guild.id,
-            role,
-          }).role);
-        })
-        .catch(reject);
+      this.rest.makeRequest('post', Constants.Endpoints.guildRoles(guild.id), true).then(role => {
+        resolve(this.rest.client.actions.GuildRoleCreate.handle({
+          guild_id: guild.id,
+          role,
+        }).role);
+      }).catch(reject);
     });
   }
 
   deleteGuildRole(role) {
     return new Promise((resolve, reject) => {
-      this.rest.makeRequest('del', Constants.Endpoints.guildRole(role.guild.id, role.id), true)
-        .then(() => {
-          resolve(this.rest.client.actions.GuildRoleDelete.handle({
-            guild_id: role.guild.id,
-            role_id: role.id,
-          }).role);
-        })
-        .catch(reject);
+      this.rest.makeRequest('del', Constants.Endpoints.guildRole(role.guild.id, role.id), true).then(() => {
+        resolve(this.rest.client.actions.GuildRoleDelete.handle({
+          guild_id: role.guild.id,
+          role_id: role.id,
+        }).role);
+      }).catch(reject);
     });
   }
 
@@ -338,23 +273,14 @@ class RESTMethods {
   getChannelMessages(channel, payload = {}) {
     return new Promise((resolve, reject) => {
       const params = [];
-      if (payload.limit) {
-        params.push(`limit=${payload.limit}`);
-      }
-      if (payload.around) {
-        params.push(`around=${payload.around}`);
-      } else if (payload.before) {
-        params.push(`before=${payload.before}`);
-      } else if (payload.after) {
-        params.push(`after=${payload.after}`);
-      }
+      if (payload.limit) params.push(`limit=${payload.limit}`);
+      if (payload.around) params.push(`around=${payload.around}`);
+      else if (payload.before) params.push(`before=${payload.before}`);
+      else if (payload.after) params.push(`after=${payload.after}`);
 
-      let request = Constants.Endpoints.channelMessages(channel.id);
-      if (params.length > 0) {
-        request += `?${params.join('&')}`;
-      }
-
-      this.rest.makeRequest('get', request, true)
+      let endpoint = Constants.Endpoints.channelMessages(channel.id);
+      if (params.length > 0) endpoint += `?${params.join('&')}`;
+      this.rest.makeRequest('get', endpoint, true)
         .then(resolve)
         .catch(reject);
     });
@@ -362,13 +288,9 @@ class RESTMethods {
 
   updateGuildMember(member, data) {
     return new Promise((resolve, reject) => {
-      if (data.channel) {
-        data.channel_id = this.rest.client.resolver.resolveChannel(data.channel).id;
-      }
+      if (data.channel) data.channel_id = this.rest.client.resolver.resolveChannel(data.channel).id;
       if (data.roles) {
-        if (data.roles instanceof Map) {
-          data.roles = data.roles.array();
-        }
+        if (data.roles instanceof Collection) data.roles = data.roles.array();
         data.roles = data.roles.map(role => role instanceof Role ? role.id : role);
       }
 
@@ -379,6 +301,7 @@ class RESTMethods {
           endpoint = Constants.Endpoints.stupidInconsistentGuildEndpoint(member.guild.id);
         }
       }
+
       this.rest.makeRequest('patch', endpoint, true, data)
         .then(resData => resolve(member.guild._updateMember(member, resData).mem))
         .catch(reject);
@@ -407,9 +330,7 @@ class RESTMethods {
   unbanGuildMember(guild, member) {
     return new Promise((resolve, reject) => {
       member = this.rest.client.resolver.resolveUser(member);
-      if (!member) {
-        throw new Error('cannot unban a user that is not a user resolvable');
-      }
+      if (!member) throw new Error('cannot unban a user that is not a user resolvable');
       const listener = (eGuild, eUser) => {
         if (guild.id === guild.id && member.id === eUser.id) {
           this.rest.client.removeListener(Constants.Events.GUILD_BAN_REMOVE, listener);
@@ -423,48 +344,30 @@ class RESTMethods {
 
   getGuildBans(guild) {
     return new Promise((resolve, reject) => {
-      this.rest.makeRequest('get', Constants.Endpoints.guildBans(guild.id), true)
-        .then(banItems => {
-          const bannedUsers = new Collection();
-          for (const banItem of banItems) {
-            const user = this.rest.client.dataManager.newUser(banItem.user);
-            bannedUsers.set(user.id, user);
-          }
-          resolve(bannedUsers);
-        })
-        .catch(reject);
+      this.rest.makeRequest('get', Constants.Endpoints.guildBans(guild.id), true).then(banItems => {
+        const bannedUsers = new Collection();
+        for (const banItem of banItems) {
+          const user = this.rest.client.dataManager.newUser(banItem.user);
+          bannedUsers.set(user.id, user);
+        }
+        resolve(bannedUsers);
+      }).catch(reject);
     });
   }
 
   updateGuildRole(role, _data) {
     return new Promise((resolve, reject) => {
-      /*
-      	can contain:
-      	name, position, permissions, color, hoist
-       */
-
       const data = {};
-
       data.name = _data.name || role.name;
       data.position = _data.position || role.position;
       data.color = _data.color || role.color;
-
-      if (data.color.startsWith('#')) {
-        data.color = parseInt(data.color.replace('#', ''), 16);
-      }
-
-      if (typeof _data.hoist !== 'undefined') {
-        data.hoist = _data.hoist;
-      } else {
-        data.hoist = role.hoist;
-      }
+      if (data.color.startsWith('#')) data.color = parseInt(data.color.replace('#', ''), 16);
+      data.hoist = typeof _data.hoist !== 'undefined' ? _data.hoist : role.hoist;
 
       if (_data.permissions) {
         let perms = 0;
         for (let perm of _data.permissions) {
-          if (typeof perm === 'string') {
-            perm = Constants.PermissionFlags[perm];
-          }
+          if (typeof perm === 'string') perm = Constants.PermissionFlags[perm];
           perms |= perm;
         }
         data.permissions = perms;
@@ -472,14 +375,12 @@ class RESTMethods {
         data.permissions = role.permissions;
       }
 
-      this.rest.makeRequest('patch', Constants.Endpoints.guildRole(role.guild.id, role.id), true, data)
-        .then(_role => {
-          resolve(this.rest.client.actions.GuildRoleUpdate.handle({
-            role: _role,
-            guild_id: role.guild.id,
-          }).updated);
-        })
-        .catch(reject);
+      this.rest.makeRequest('patch', Constants.Endpoints.guildRole(role.guild.id, role.id), true, data).then(_role => {
+        resolve(this.rest.client.actions.GuildRoleUpdate.handle({
+          role: _role,
+          guild_id: role.guild.id,
+        }).updated);
+      }).catch(reject);
     });
   }
 
@@ -526,16 +427,14 @@ class RESTMethods {
 
   getGuildInvites(guild) {
     return new Promise((resolve, reject) => {
-      this.rest.makeRequest('get', Constants.Endpoints.guildInvites(guild.id), true)
-        .then(inviteItems => {
-          const invites = new Collection();
-          for (const inviteItem of inviteItems) {
-            const invite = new Invite(this.rest.client, inviteItem);
-            invites.set(invite.code, invite);
-          }
-          resolve(invites);
-        })
-        .catch(reject);
+      this.rest.makeRequest('get', Constants.Endpoints.guildInvites(guild.id), true).then(inviteItems => {
+        const invites = new Collection();
+        for (const inviteItem of inviteItems) {
+          const invite = new Invite(this.rest.client, inviteItem);
+          invites.set(invite.code, invite);
+        }
+        resolve(invites);
+      }).catch(reject);
     });
   }
 }
